added Schema from file factory; refactoring
